Clarify how Screen classes work
- Add some docs explaining architecture
- Add an abstract Screen class for each screen to extend
- Rename all screen classes to have a "Screen" suffix
